requirements.txt
ultralytics
streamlit
opencv-python
numpy
Pillow
